# Conflicts:
#	upload/admin_area/main.php
#	upload/cb_install/sql/configs.sql
#	upload/js/functions.min.js
#	upload/styles/cb_28/theme/css/all.min.css

+++ upload/manage_videos.php
@@ -8,18 +8,19 @@ if( config('videosSection') != 'yes' ){
     redirect_to(BASEURL);
 }
 
-global $userquery, $cbvideo, $eh, $pages, $cbvid;
+global $cbvideo, $eh, $pages, $cbvid;
 
-$userquery->logincheck();
-$udetails = $userquery->get_user_details(user_id());
+userquery::getInstance()->logincheck();
+$udetails = userquery::getInstance()->get_user_details(user_id());
 assign('user', $udetails);
-assign('p', $userquery->get_user_profile($udetails['userid']));
+assign('p', userquery::getInstance()->get_user_profile($udetails['userid']));
 
 $mode = $_GET['mode'];
 
 $page = mysql_clean($_GET['page']);
 $get_limit = create_query_limit($page, config('videos_list_per_page'));
-
+$favorites = User::getInstance()->getFavoritesVideos(user_id());
+assign('favorites', $favorites);
 assign('queryString', queryString(null, ['type',
     'vid_delete']));
 switch ($mode) {
@@ -29,14 +30,18 @@ switch ($mode) {
 
         //Deleting Video
         if (!empty($_GET['vid_delete'])) {
-            $video = mysql_clean($_GET['vid_delete']);
-            $cbvideo->delete_video($video);
+            $videoid = mysql_clean($_GET['vid_delete']);
+            if ($cbvid->is_video_owner($videoid, user_id())) {
+                $cbvideo->delete_video($videoid);
+            }
         }
 
         //Deleting Videos
         if (isset($_POST['delete_videos'])) {
             for ($id = 0; $id <= config('videos_list_per_page'); $id++) {
-                $cbvideo->delete_video($_POST['check_vid'][$id]);
+                if ($cbvid->is_video_owner($_POST['check_vid'][$id], user_id())) {
+                    $cbvideo->delete_video($_POST['check_vid'][$id]);
+                }
             }
             $eh->flush();
             e(lang('vdo_multi_del_erro'), 'm');
@@ -56,6 +61,7 @@ switch ($mode) {
         //Collecting Data for Pagination
         $vid_array['count_only'] = true;
         $total_rows = get_videos($vid_array);
+
         $total_pages = count_pages($total_rows, config('videos_list_per_page'));
 
         //Pagination
@@ -68,18 +74,9 @@ switch ($mode) {
     case 'favorites':
         assign('mode', 'favorites');
         //Removing video from favorites
-        if (!empty($_GET['vid_delete'])) {
-            $video = mysql_clean($_GET['vid_delete']);
-            $cbvideo->action->remove_favorite($video);
-        }
-
-        //Removing Multiple Videos
-        if (isset($_POST['delete_fav_videos'])) {
-            for ($id = 0; $id <= config('videos_list_per_page'); $id++) {
-                $cbvideo->delete_video($_POST['check_vid'][$id]);
-            }
-            $eh->flush();
-            e(lang('vdo_multi_del_fav_msg'), 'm');
+        if (!empty($_GET['remove_fav_videoid']) && in_array($_GET['remove_fav_videoid'], $favorites) ) {
+            $videoid = mysql_clean($_GET['remove_fav_videoid']);
+            $cbvideo->action->remove_favorite($videoid);
         }
         if (get('query') != '') {
             $cond = " (video.title LIKE '%" . mysql_clean(get('query')) . "%' OR video.tags LIKE '%" . mysql_clean(get('query')) . "%' )";
